% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdocsyntax.r
\name{unhighlight_html_docs}
\alias{unhighlight_html_docs}
\title{Disable HTML documentation syntax highlighting}
\usage{
unhighlight_html_docs()
}
\description{
Revert to the original help server for handling HTML documentation.
}
\seealso{
\code{\link{highlight_html_docs}}
}
